lolbin-detection-system/tray-app/requirements.txt
pystray==0.19.5
pillow==10.0.0
websocket-client==1.7.0
requests==2.31.0